Move padding to `summary` for larger tap target
When `details` is closed, the padding on this element makes the box larger than the tap target, since you can only click on `summary` itself. By moving the padding to `summary` when `details` is collapsed, the tap target takes up the entire visual box. Then it can be moved back to `details` when expanded to keep padding around the contents.

I suppose it's also possible to use `padding: unset` or similar for `details[open]` if you wanted to keep it more general than `padding: 0;`

Love this simple library - I recognize a lot has gone into it. Thanks for maintaining!